Updated astali test folder structure and mocks generation.
