--- test/settings_menu_eh_test.mocks.dart
@@ -1,34 +1,0 @@
-// Mocks generated by Mockito 5.3.0 from annotations
-// in astali/test/settings_menu_eh_test.dart.
-// Do not manually edit this file.
-
-// ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'package:astali/settings/settings_menu_spawner.dart' as _i2;
-import 'package:flutter/material.dart' as _i3;
-import 'package:mockito/mockito.dart' as _i1;
-
-// ignore_for_file: type=lint
-// ignore_for_file: avoid_redundant_argument_values
-// ignore_for_file: avoid_setters_without_getters
-// ignore_for_file: comment_references
-// ignore_for_file: implementation_imports
-// ignore_for_file: invalid_use_of_visible_for_testing_member
-// ignore_for_file: prefer_const_constructors
-// ignore_for_file: unnecessary_parenthesis
-// ignore_for_file: camel_case_types
-// ignore_for_file: subtype_of_sealed_class
-
-/// A class which mocks [SettingsMenuSpawner].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockSettingsMenuSpawner extends _i1.Mock
-    implements _i2.SettingsMenuSpawner {
-  MockSettingsMenuSpawner() {
-    _i1.throwOnMissingStub(this);
-  }
-
-  @override
-  void spawnSettingsMenu(_i3.BuildContext? context) =>
-      super.noSuchMethod(Invocation.method(#spawnSettingsMenu, [context]),
-          returnValueForMissingStub: null);
-}
