+++ test/test_doubles/generated_mocks.dart
@@ -1,0 +1,12 @@
+// Copyright (C) 2022 Andrea Ballestrazzi
+
+// Astali
+import 'package:astali/settings/settings_menu_spawner.dart';
+import 'package:astali/settings/default_settings_menu_spawner.dart';
+
+// Mockito
+import 'package:mockito/annotations.dart';
+
+@GenerateMocks([SettingsMenuSpawner])
+@GenerateMocks([DefaultSettingsMenuEventHandler])
+void main() {}
