+++ test/test_doubles/generated_mocks.mocks.dart
@@ -1,9 +1,11 @@
 // Mocks generated by Mockito 5.3.0 from annotations
-// in astali/test/settings_menu_widget_test.dart.
+// in astali/test/test_doubles/generated_mocks.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'package:astali/settings/default_settings_menu_spawner.dart' as _i2;
+import 'package:astali/settings/default_settings_menu_spawner.dart' as _i4;
+import 'package:astali/settings/settings_menu_spawner.dart' as _i2;
+import 'package:flutter/material.dart' as _i3;
 import 'package:mockito/mockito.dart' as _i1;
 
 // ignore_for_file: type=lint
@@ -17,11 +19,26 @@ import 'package:mockito/mockito.dart' as _i1;
 // ignore_for_file: camel_case_types
 // ignore_for_file: subtype_of_sealed_class
 
+/// A class which mocks [SettingsMenuSpawner].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockSettingsMenuSpawner extends _i1.Mock
+    implements _i2.SettingsMenuSpawner {
+  MockSettingsMenuSpawner() {
+    _i1.throwOnMissingStub(this);
+  }
+
+  @override
+  void spawnSettingsMenu(_i3.BuildContext? context) =>
+      super.noSuchMethod(Invocation.method(#spawnSettingsMenu, [context]),
+          returnValueForMissingStub: null);
+}
+
 /// A class which mocks [DefaultSettingsMenuEventHandler].
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockDefaultSettingsMenuEventHandler extends _i1.Mock
-    implements _i2.DefaultSettingsMenuEventHandler {
+    implements _i4.DefaultSettingsMenuEventHandler {
   MockDefaultSettingsMenuEventHandler() {
     _i1.throwOnMissingStub(this);
   }
